kafka-server/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.0
mysql-connector-python==8.2.0
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0